Rewrite traverseTree to iterative BFS for scale.
When the Prometheus plugin is used on a Jenkins installation with a large number of builds and metrics,
it hits the inherent limitations of a recursive algorithm and crashes with a stack overflow. Recursive
algorithms are not ideal in production because of this, so rewrite this to be iterative.

Emperical testing has shown that this rewrite fixes the issue and allows Jenkins installations to work and metrics to be
scraped properly from Prometheus.

1. Use a standard iterative BFS algorithm.
2. Remove TreeMap for collision purposes (The O(log(n)) sorting of the keys is unncessary) and replace with HashSet.

Addresses:
- https://github.com/jenkinsci/prometheus-plugin/issues/127
- https://github.com/jenkinsci/prometheus-plugin/issues/128
- https://github.com/jenkinsci/prometheus-plugin/issues/134
- https://issues.jenkins-ci.org/browse/JENKINS-59966