ISSUE-204 - last build metrics don't have stages breakdown
Add processing of the last completed job, which, we also break down to stages.
Took the opportunity to unify code between processing a generic run when iterating everything, and when processing the last completed one.
Also, added a subclass to handle Run specific metric collectors to make it easier to add more metrics in the future and reduce code duplication