Don't export disk usage metrics of orphaned items
Before this change, the gauges for Jenkins disk usage were cached
forever. The disk usage plugin prunes old items from time to time, but
the metrics previously collected into the gauges are not affected in any
way by this. The easiest way not to export no-longer-existing items is
to just recreate the gauges every time for every call to collect().