Add metrics about file store usage
In addition to the storage statistics for folders and jobs, also add
metrics about file stores (a.k.a. disks / partitions) on which those
folders and jobs are stored.

This adds the two new metrics file_store_capacity_bytes and
file_store_available_bytes, along with the new label file_store, that's
attached to each folder and job and identifies on which file store they
live.