disable DiskUsageCollector when its option is unchecked (JENKINS-66790)
The `DiskUsageCollector` class uses the `ConfigurationUtils#getCollectDiskUsage` method to determine whether the it should be enabled or not. Unfortunately, the mentioned method ignores the checkbox configuration option available on the `Configure System` page. The `PrometheusConfiguration` class keeps the requested value in the `collectDiskUsage` field. It is calculated based on the saved configuration which fallback to the `COLLECT_DISK_USAGE` environment variable or `true` (if the environment variable is not set). The `ConfigurationUtils#getCollectDiskUsage` method is updated to execute the `PrometheusConfiguration#getCollectDiskUsage()` method, so the user configuration is taken into account.