Disable DiskUsageCollector when its option is unchecked (JENKINS-66790)
The `DiskUsageCollector` class uses the `ConfigurationUtils#getCollectDiskUsage` method to determine whether the disk usage statistics should be collected or not. Unfortunately, the mentioned method ignores the checkbox configuration option available on the `Configure System` page. The `PrometheusConfiguration` class keeps the requested value in the `collectDiskUsage` field which is set with the default value if needed. The `ConfigurationUtils#getCollectDiskUsage` method is removed and the `PrometheusConfiguration#getCollectDiskUsage` method is used instead.

The `PrometheusConfiguration#getDefaultCollectDiskUsage` method is no longer used, so it has been removed too.