Use Mockito Inline instead of PowerMock and cleanup dependencies
The project build was broken because the PowerMock libraries version has not been set. However, PowerMock is not compatible with Mockito 3.X, so there is not sense to fix it. Instead the Mockito Inline library is added which supports mocking static methods. The tests were updated to use the new API.

The Jenkins version has been updated to the latest available LTS. Every LTS is provided with a `bom-X.Y.Z` artifact which should be used to determine versions of the most common dependencies (e.g. the `junit` plugin). Whenever the LTS is updated to a newer version, the `bom` artifact must be updated too. All unnecessary dependencies have been removed.

The `Jenkins#getInstance` method is deprecated, so all occurrences have been replaced with `Jenkins#get`.